Split context management into API controlled and ad-hoc
Split initialization into two modes, depending on whether the main
extension is present.

If the main extension exists, the language handler will wait for a
configuration call through the API before parsing or setting any
configuration. In this mode, the extension will not store or load from
the workspace settings, but rely fully on the API setting. It also will
not change the configuration file set based on the open editor.

If the main extension doesn't exist, the old initialization procedure
will be performed, and the extension will be blocked until west and the
Zephyr base is found.