Language Server (#11)
This is a full refactor of the Kconfig extension, moving all .conf file and workspace level language support to a Python based language server, built on top of Zephyr's kconfiglib. 

kconfiglib has no way to perform a partial reparse of Kconfig files, and is just generally too slow for live feedback when editing Kconfig files. To get around this, the Typescript part of the extension still contains a parser that only operates on single Kconfig files, and provides local syntax checking and language support.

The overall architecture is described in ARCHITECTURE.md.